Update form keys
The form keys are cased incorrectly. While it could make sense
to just case them correctly on the client side, it seems to me at least
that it's easier to assume all variables will be camelcased (which in
this case has the effect of making them all lowercase).

Use list assignment to pull the values out of the form data once.

Update the tests to reflect this as well.